apply offset to each node
Apply offset to each node when generating DEA file.

To generate DAE that can be smoothly imported to
Unity game engine, create DAE with small absolute
values for coordinates.

To achive that, we apply the offset to either node's
mesh vertices or node's transfrom matrix.

We use simple heuristic to decide if mesh vertices or
transform should be changed. The heuristic seems to work
good with IFC generated by Revit, Tekla Structures and
Carlson CAD Software.